refactor: architecture review P0-P3 — type safety, RAII, and bug fixes
P0: Timer::type int → enum class TimerType (15 files)
P1: MXP maps raw new/qDeleteAll → std::unique_ptr with std::map/vector
P2a: 6 unscoped enums → enum class (Phase, ActionSource, HistoryStatus,
     ScriptReloadOption, AutoConnect, MXPMode)
P2b: send_to → SendTo enum class, colour_change_type → ColourChangeType
P2c: NotepadWidget QList → QPointer, view pointers documented
P2d: LuaDatabase copy/move deleted
P2e: z_stream wrapped in RAII ZlibStream guard
P3a: SoundManager QAudioEngine* → unique_ptr
P3b: Fixed scriptFileWatcher double-free, removed redundant manual deletes
P3c: Fixed SoundManager async lambda use-after-free (shared_ptr alive guard)
P3d: Fixed timer offset>=interval busy-loop and fractional second truncation
P3e: 25+ #define constants → inline constexpr
P3f: executeLua long& → qint32& (LP64 narrowing fix)
P3g: queue() → std::expected<void, WorldError>

749/749 tests pass.

+++ src/world/lua_api/world_settings.cpp
@@ -1181,7 +1181,7 @@ int L_GetInfo(lua_State* L)
             break;
 
         case 239: // Current action source (trigger/alias/timer/etc)
-            lua_pushinteger(L, pDoc->m_iCurrentActionSource);
+            lua_pushinteger(L, static_cast<quint32>(pDoc->m_iCurrentActionSource));
             break;
 
         case 240: // Average character width
